% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addggText}
\alias{addggText}
\title{Assign text annotation to the ggplot scatterplot}
\usage{
addggText(plotData, vars, dims)
}
\arguments{
\item{plotData}{data.frame. The data to use for plotting.}

\item{vars}{character. The variables whose levels will be depicted.}

\item{dims}{character. The dimensions to be plotted.}
}
\value{
A vector of names
}
\description{
Helper function
}
\keyword{addggText}
