% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DataTester}
\alias{DataTester}
\title{Modify the data if the subsequent steps of DIMER have been used}
\usage{
DataTester(Data, func)
}
\arguments{
\item{Data}{list. The outcome of readData().}

\item{func}{character. Which function is used.}
}
\value{
The Data list after the modification (removal of some slots)
}
\description{
Helper function
}
\keyword{Sample.DataTester}
