% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DateStamp}
\alias{DateStamp}
\title{Create a date stamp}
\usage{
DateStamp(date)
}
\arguments{
\item{date}{character. The date.}
}
\value{
A date to use in the filename
}
\description{
Helper function
}
\keyword{DateStamp}
