% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DEplot_entries}
\alias{DEplot_entries}
\title{Find the marker of interest in the differential expression results}
\usage{
DEplot_entries(deData, filters)
}
\arguments{
\item{deData}{matrix. The differential expression analysis results.}

\item{filters}{character. The filtering variables.}
}
\value{
A vector of indices
}
\description{
Helper function
}
\keyword{DEplot_entries}
