% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsGenerator.R
\name{doMA}
\alias{doMA}
\title{Generates the MA plot}
\usage{
doMA(
  Data,
  filters,
  logFC = 1,
  FDR = 0.01,
  show.more = FALSE,
  size = 5,
  show.plot = TRUE,
  save.plot = FALSE,
  save.table = FALSE,
  interactive = TRUE
)
}
\arguments{
\item{Data}{list. The outcome of MarkerQuery().}

\item{filters}{character. A set of conditions that specifies which differential expression results should be plotted. It should ALWAYS be a level of the Comparison variable of the DEstats component.}

\item{logFC}{numeric. A positive value specifying the symmetric logFC cut-off to be depicted. Default is 1.}

\item{FDR}{numeric. A value between (0,1] specifying the FDR cutoff to be depicted. Default is 0.01.}

\item{show.more}{logical. If TRUE, all markers of Data$filteredData$Marker_List are highlighted, otherwise only the marker of Data$selectedData$Marker_List is highlighted. Default is FALSE.}

\item{size}{numeric. It specifies the plotted dot size. Default is 5.}

\item{show.plot}{logical. If TRUE, it shows the plot on the screen. Default is TRUE.}

\item{save.plot}{logical. If TRUE, it saves the plot in an automatically generated folder <data folder>/Output/<InteractivePlots or NonInteractivePlots>. Default is FALSE.}

\item{save.table}{logical. If TRUE, it saves the plot in an automatically generated folder <data folder>/Output/Tables. Default if FALSE.}

\item{interactive}{logical. If TRUE, it generates an interactive plot with plotly otherwise a static plot with ggplot. Default is TRUE.}
}
\value{
data list. The plots and the tables of interest stored in automatically generated <data folder>/Output/ folders. The tables include a summary table of quantiles of the marker expression, the differential expression analysis results of the markers of interest and the normalised data for the markers of interest.
}
\description{
It creates the MA plot for a particular cell type and condition
}
\examples{
data(nMyo_Data)
data<-readData(nMyo_Data,Markers_file=NULL,is.Exact=TRUE,logFC_cutoff=0,FDR_cutoff=1)

data<-MarkerQuery(Data=data,marker="Postn")

data<-doMA(Data=data,
         filters="Comparison=='Fibroblast: Sham-MI'",
         logFC=1,
         FDR=0.01,
         show.plot=TRUE,
         save.plot=FALSE)

}
\keyword{doMA}
