% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsGenerator.R
\name{doScatterDR}
\alias{doScatterDR}
\title{Generates the dimensionality reduction scatterplot}
\usage{
doScatterDR(
  Data,
  grouping.by = NULL,
  highlight.by = "CellType",
  showDims = c("Dim1", "Dim2"),
  size = 5,
  show.plot = TRUE,
  save.plot = FALSE,
  save.table = FALSE,
  interactive = TRUE
)
}
\arguments{
\item{Data}{list. The outcome of MarkerQuery().}

\item{grouping.by}{character. A grouping variable to be depicted in the plot. It groups (color code) the samples with common values for this variable. It can take any of the factor variables depicted at the column names of Data$Design. Default is NULL.}

\item{highlight.by}{character. A second grouping variable to highlight certain point on the INTERACTIVE plot upon double click. It highlights the samples with common values for this variable and shades the rest. It can take any of the factor variables depicted at the column names of Data$Design. Default is CellType.}

\item{showDims}{character vector. A character vector of size 2 or 3 specifing the dimensions to be plotted (Dim1, Dim2 and Dim3). If all three are specified then a 3D plot is generated otherwise a 2D plot of the specified dimensions is returned. Default is c("Dim1","Dim2"), i.e. the first 2 dimensionality reduction components.}

\item{size}{numeric. It specifies the plotted dot size. Default is 5.}

\item{show.plot}{logical. If TRUE, it shows the plot on the screen. Default is TRUE.}

\item{save.plot}{logical. If TRUE, it saves the plot in an automatically generated folder <data folder>/Output/<InteractivePlots or NonInteractivePlots>. Default is FALSE.}

\item{save.table}{logical. If TRUE, it saves the plot in an automatically generated folder <data folder>/Output/Tables. Default if FALSE.}

\item{interactive}{logical. If TRUE, it generates an interactive plot with plotly otherwise a static plot with ggplot. Default is TRUE.}
}
\value{
data list. The plots and the tables of interest stored in automatically generated <data folder>/Output/ folders. The tables include a summary table of quantiles of the marker expression, the differential expression analysis results of the markers of interest and the normalised data for the markers of interest.
}
\description{
It creates a dimensionality reduction plot (e.g. PCA or t-SNE) highlighting the samples of interest
}
\examples{
data(nMyo_Data)
data<-readData(nMyo_Data,Markers_file=NULL,is.Exact=TRUE,logFC_cutoff=0,FDR_cutoff=1)

data<-MarkerQuery(Data=data,marker="Postn")

data<-doScatterDR(Data=data,
         highlight.by="CellType",
         show.plot=TRUE,
         save.plot=FALSE)

}
\keyword{doScatterDR}
