% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{filterSampleData}
\alias{filterSampleData}
\title{Filter the Samples of the Design file based on the filters}
\usage{
filterSampleData(designData, filters)
}
\arguments{
\item{designData}{data.frame. The data of the Design_file.}

\item{filters}{expression. The filters.}
}
\value{
The filtered data of the Design_file
}
\description{
Helper function
}
\keyword{filterSampleData}
