% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{filterSampleDEstatsData}
\alias{filterSampleDEstatsData}
\title{Filter the Samples of the DE file based on the filters}
\usage{
filterSampleDEstatsData(Data)
}
\arguments{
\item{Data}{list. The outcome of the readData().}
}
\value{
The filtered data of the DE_file
}
\description{
Helper function
}
\keyword{filterSampleDEstatsData}
