% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{finaliseMarkerList}
\alias{finaliseMarkerList}
\title{Read the DE_file and tests the data validity}
\usage{
finaliseMarkerList(data, genes)
}
\arguments{
\item{data}{list. The outcome of the RequiredDataReader() internal function updated with other components (is.Exact, logFC_cutoff, FDR_cutoff, status).}

\item{genes}{character. A list of genes.}
}
\value{
A list with the significant DE genes, the final list of genes and the missing genes
}
\description{
Helper function
}
\keyword{finaliseMarkerList}
