% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{findGeneType}
\alias{findGeneType}
\title{Find the marker of interest in the data}
\usage{
findGeneType(templ, genes, multiple = FALSE)
}
\arguments{
\item{templ}{matrix. It contains the marker annotations.}

\item{genes}{character. The IDs of the markers of interest.}

\item{multiple}{logical. Whether the temp contains repeats of the same gene name. Default is FALSE.}
}
\value{
A vector of indices
}
\description{
Helper function
}
\keyword{findGeneType}
