% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{firstStop}
\alias{firstStop}
\title{Stop the analysis if the files are not provided or the filenames are wrong}
\usage{
firstStop(Counts_file, Design_file)
}
\arguments{
\item{Counts_file}{character. The location of the normalised marker expression file.}

\item{Design_file}{character. The location of the experimental design file.}
}
\value{
A stop or nothing
}
\description{
Helper function
}
\keyword{firstStop}
