% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fixFilteredData}
\alias{fixFilteredData}
\title{It creates the filteredData slot}
\usage{
fixFilteredData(data, sigDE.genes.status)
}
\arguments{
\item{data}{list. The outcome of the RequiredDataReader() internal function updated with other components (is.Exact, logFC_cutoff, FDR_cutoff, status,Output_folder and Dimensions).}

\item{sigDE.genes.status}{list. The outcome of finaliseMarkerList().}
}
\value{
A list with the filteredData
}
\description{
Helper function
}
\keyword{fixFilteredData}
