% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GeneNames}
\alias{GeneNames}
\title{GeneNames}
\usage{
GeneNames(data, symbol = ":", what = "Name")
}
\arguments{
\item{data}{character vector. A vector of IDs. Each entry may contain two ID separated by a symbol e.g. EnsemblID:MarkerName.}

\item{symbol}{character. The separator used. Default is the symbol ':'.}

\item{what}{character. Which of the above two IDs (if two) to keep. Options are 'Name' (default) to keep the marker name or 'Other' to keep the other ID. If each entry consists of one ID then this is the one kept.}
}
\value{
A vector of IDs
}
\description{
Extract gene IDs
}
\keyword{GeneNames}
