% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Splits the violin into two components}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{...}{other arguments}
}
\value{
A data.frame
}
\description{
Helper function
}
\keyword{geom_split_violin}
