% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Color specifications}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{numeric. The number of colors to be used from the pallete.}
}
\value{
A vector of colors
}
\description{
color_hue
}
\keyword{gg_color_hue}
