% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{\%&\%}
\alias{\%&\%}
\title{Join two strings}
\usage{
x \%&\% y
}
\arguments{
\item{x}{string. The first string.}

\item{y}{string. The second string.}
}
\value{
A joined string
}
\description{
Helper function
}
\keyword{joinstring}
