% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makecolors_scatterplot}
\alias{makecolors_scatterplot}
\title{Assign colors to the scatterplot}
\usage{
makecolors_scatterplot(Data, plotData, feature)
}
\arguments{
\item{Data}{list. The outcome of the MarkerQuery().}

\item{plotData}{data.frame. The data to use for plotting.}

\item{feature}{character. The feature to be plotted.}
}
\value{
A list
}
\description{
Helper function
}
\keyword{makecolors_scatterplot}
