% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makecolors_violin}
\alias{makecolors_violin}
\title{Assign colors to the violin}
\usage{
makecolors_violin(plotData, feature, grouping.by, split.by)
}
\arguments{
\item{plotData}{list. The outcome of the MarkerQuery().}

\item{feature}{character. The feature to be plotted.}

\item{grouping.by}{character.  The grouping variable to be highlighted in the x-axis.}

\item{split.by}{character. The variable to split each violin with.}
}
\value{
A list
}
\description{
Helper function
}
\keyword{makecolors_violin}
