% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerQuery.R
\name{MarkerQuery}
\alias{MarkerQuery}
\title{Select the marker of interest for visualisation}
\usage{
MarkerQuery(Data, marker)
}
\arguments{
\item{Data}{list. The outcome of readData().}

\item{marker}{character. The ID of the marker to be visualised.}
}
\value{
data list. A list includying the outcome of readData() and an extra slot 'selectedData' with the data of the marker that has been selected for visualisation.
}
\description{
It selects the marker to be plotted and tabulated.
}
\examples{
data(nMyo_Data)
data<-readData(nMyo_Data,Markers_file=NULL,is.Exact=TRUE,logFC_cutoff=0,FDR_cutoff=1)
data<-MarkerQuery(Data=data,marker="Postn")

}
\keyword{MarkerQuery}
