% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{matchDEentries}
\alias{matchDEentries}
\title{Find the marker of interest in the differential expression results}
\usage{
matchDEentries(deEntry, the_other, the_character)
}
\arguments{
\item{deEntry}{character. A marker characteristics from the DE file.}

\item{the_other}{character. A marker characteristic from the design file.}

\item{the_character}{character. A character to split the marker characteristics from the DE file.}
}
\value{
A vector of indices
}
\description{
Helper function
}
\keyword{matchDEentries}
