\name{nMyo_Data}
\alias{nMyo_Data}
\docType{data}
\title{
nMyo_Data
}
\description{
nMyo_Data
}
\usage{data("nMyo_Data")}
\format{
  The format is:
List of 3
 $ CorrCounts :'data.frame':	26942 obs. of  2032 variables:
  ..$ Marker : chr [1:26942] "ENSMUSG00000000001:Gnai3" "ENSMUSG00000000028:Cdc45" "ENSMUSG00000000031:H19" "ENSMUSG00000000037:Scml2" ...
  ..$ RMH2238: num [1:26942] 12.9 0 0 0 0 ...
  ..$ RMH2239: num [1:26942] 10.9 0 0 0 0 ...
  ..$ RMH2240: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2241: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2244: num [1:26942] 4.68 0 0 0 0 ...
  ..$ RMH2245: num [1:26942] 11.57 0 8.35 0 0 ...
  ..$ RMH2246: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2247: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2249: num [1:26942] 0 0 0 0 4.72 ...
  ..$ RMH2250: num [1:26942] 5.73 0 0 0 4.67 ...
  ..$ RMH2251: num [1:26942] 13.1 0 0 0 0 ...
  ..$ RMH2252: num [1:26942] 8.67 0 0 0 0 ...
  ..$ RMH2254: num [1:26942] 10.8 0 0 0 0 ...
  ..$ RMH2255: num [1:26942] 5.38 0 0 0 9.35 ...
  ..$ RMH2256: num [1:26942] 5.93 0 0 0 0 ...
  ..$ RMH2257: num [1:26942] 10.6 0 0 0 0 ...
  ..$ RMH2258: num [1:26942] 9.91 0 0 0 0 ...
  ..$ RMH2260: num [1:26942] 11.1 0 0 0 0 ...
  ..$ RMH2261: num [1:26942] 4.73 0 0 0 0 ...
  ..$ RMH2262: num [1:26942] 11.5 0 0 0 0 ...
  ..$ RMH2263: num [1:26942] 5.48 0 0 0 0 ...
  ..$ RMH2264: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2265: num [1:26942] 0 5.07 0 0 0 ...
  ..$ RMH2266: num [1:26942] 12.8 0 0 0 0 ...
  ..$ RMH2267: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2268: num [1:26942] 7.6 11.7 0 0 0 ...
  ..$ RMH2270: num [1:26942] 12.42 7.29 0 0 0 ...
  ..$ RMH2273: num [1:26942] 11.3 0 0 0 0 ...
  ..$ RMH2274: num [1:26942] 11.4 0 0 0 0 ...
  ..$ RMH2276: num [1:26942] 11 0 0 0 0 ...
  ..$ RMH2277: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2278: num [1:26942] 5.98 0 0 0 0 ...
  ..$ RMH2279: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2281: num [1:26942] 11.9 0 0 0 0 ...
  ..$ RMH2282: num [1:26942] 5.49 0 0 0 0 ...
  ..$ RMH2283: num [1:26942] 10.9 0 0 0 0 ...
  ..$ RMH2285: num [1:26942] 10.2 0 0 0 0 ...
  ..$ RMH2286: num [1:26942] 5.67 0 0 0 5.01 ...
  ..$ RMH2287: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2288: num [1:26942] 0 0 0 0 12 ...
  ..$ RMH2289: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2290: num [1:26942] 8.32 0 0 0 0 ...
  ..$ RMH2291: num [1:26942] 10.82 0 0 0 4.76 ...
  ..$ RMH2292: num [1:26942] 11.73 0 0 0 4.52 ...
  ..$ RMH2293: num [1:26942] 0 0 0 0 10.7 ...
  ..$ RMH2294: num [1:26942] 10.2 0 0 0 0 ...
  ..$ RMH2295: num [1:26942] 0 0 0 0 6.05 ...
  ..$ RMH2296: num [1:26942] 11.8 0 0 0 0 ...
  ..$ RMH2297: num [1:26942] 10.7 0 0 10.2 10.5 ...
  ..$ RMH2298: num [1:26942] 9.56 0 0 0 0 ...
  ..$ RMH2300: num [1:26942] 9.97 0 0 0 0 ...
  ..$ RMH2301: num [1:26942] 12 0 0 0 0 ...
  ..$ RMH2302: num [1:26942] 11.9 0 0 0 0 ...
  ..$ RMH2303: num [1:26942] 11.6 0 0 0 0 ...
  ..$ RMH2305: num [1:26942] 6.68 0 0 0 0 ...
  ..$ RMH2306: num [1:26942] 4.99 0 0 0 0 ...
  ..$ RMH2308: num [1:26942] 10.3 0 0 0 0 ...
  ..$ RMH2309: num [1:26942] 0 0 0 0 8.07 ...
  ..$ RMH2310: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2311: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2312: num [1:26942] 0 0 0 0 5.31 ...
  ..$ RMH2314: num [1:26942] 9.77 0 0 0 0 ...
  ..$ RMH2316: num [1:26942] 5.54 0 0 0 0 ...
  ..$ RMH2317: num [1:26942] 12.8 0 0 0 0 ...
  ..$ RMH2318: num [1:26942] 10.6 0 0 0 0 ...
  ..$ RMH2321: num [1:26942] 5.68 0 0 0 0 ...
  ..$ RMH2322: num [1:26942] 9.13 0 0 0 11.15 ...
  ..$ RMH2326: num [1:26942] 4.62 0 0 0 0 ...
  ..$ RMH2327: num [1:26942] 12.6 0 0 0 0 ...
  ..$ RMH2330: num [1:26942] 5.6 0 0 0 0 ...
  ..$ RMH2331: num [1:26942] 5.66 0 0 0 0 ...
  ..$ RMH2333: num [1:26942] 8.07 0 0 0 0 ...
  ..$ RMH2334: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2335: num [1:26942] 13.4 0 0 0 0 ...
  ..$ RMH2336: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2337: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2338: num [1:26942] 5.89 0 0 0 0 ...
  ..$ RMH2339: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2340: num [1:26942] 13.1 0 0 0 0 ...
  ..$ RMH2342: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2343: num [1:26942] 9.83 0 0 0 0 ...
  ..$ RMH2344: num [1:26942] 11.82 0 0 0 7.88 ...
  ..$ RMH2345: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2347: num [1:26942] 5.68 0 0 0 6.51 ...
  ..$ RMH2348: num [1:26942] 0 0 0 0 10.4 ...
  ..$ RMH2349: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2350: num [1:26942] 5.76 0 0 0 0 ...
  ..$ RMH2351: num [1:26942] 6.47 0 9.32 0 0 ...
  ..$ RMH2352: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2353: num [1:26942] 12.3 0 0 0 0 ...
  ..$ RMH2354: num [1:26942] 13.6 0 0 0 0 ...
  ..$ RMH2355: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2356: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2357: num [1:26942] 5.71 0 0 0 0 ...
  ..$ RMH2358: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2359: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2360: num [1:26942] 0 0 0 0 0 ...
  ..$ RMH2361: num [1:26942] 0 0 0 0 0 ...
  .. [list output truncated]
 $ DesignTable:'data.frame':	2031 obs. of  23 variables:
  ..$ SampleID        : chr [1:2031] "RMH2238" "RMH2239" "RMH2240" "RMH2241" ...
  ..$ Plate           : chr [1:2031] "P3" "P3" "P3" "P3" ...
  ..$ Mouse           : chr [1:2031] "A-Sham" "A-Sham" "A-Sham" "A-Sham" ...
  ..$ Condition       : chr [1:2031] "Sham" "Sham" "Sham" "Sham" ...
  ..$ LibrarySize     : int [1:2031] 2833319 1934676 2395588 1964128 1826143 1052373 1868916 1725347 2147195 2492999 ...
  ..$ Spike_n         : int [1:2031] 14423 3074 20208 4306 5780 1555 38021 4169 5804 23447 ...
  ..$ Spike_p         : num [1:2031] 0.47 0.147 0.769 0.203 0.293 ...
  ..$ Ribosomal_p     : num [1:2031] 6.45 6.69 7.68 6.8 6.58 ...
  ..$ Mitochondrial_p : num [1:2031] 0.773 0.586 0.409 0.365 0.563 ...
  ..$ Detected_genes  : int [1:2031] 3331 3934 3248 3795 3581 4101 3532 3322 3316 3061 ...
  ..$ Batch           : chr [1:2031] "B1" "B1" "B1" "B1" ...
  ..$ CycloneGroups   : chr [1:2031] "G1" "G1" "G1" "G1" ...
  ..$ Dim1            : num [1:2031] 5.52 -2.58 -2.26 -1.91 -5.2 ...
  ..$ Dim2            : num [1:2031] -0.0217 -2.2576 -2.0803 -1.4918 5.7249 ...
  ..$ Dim3            : num [1:2031] 1.022 -0.623 -1.506 -4.108 2.597 ...
  ..$ AP_clusters     : chr [1:2031] "AP5" "AP9" "AP29" "AP1" ...
  ..$ Louvain_clusters: chr [1:2031] "L7" "L1" "L0" "L1" ...
  ..$ Jaccard         : num [1:2031] 0.797 0.615 0.61 0.772 0.968 ...
  ..$ AP_type         : chr [1:2031] "Stable" "Pattern" "Pattern" "Stable" ...
  ..$ CellType        : chr [1:2031] "Endothelial" "Fibroblast" "Fibroblast" "Fibroblast" ...
  ..$ Seurat_groups   : chr [1:2031] "Endothelial" "Fibroblast" "Fibroblast" "Fibroblast" ...
  ..$ Colors_CellType : chr [1:2031] "#F8766D" "#00B0F6" "#00B0F6" "#00B0F6" ...
  ..$ Colors_Condition: chr [1:2031] "#00B0F6" "#00B0F6" "#00B0F6" "#00B0F6" ...
 $ DEs        :'data.frame':	646608 obs. of  7 variables:
  ..$ Marker    : chr [1:646608] "ENSMUSG00000000001:Gnai3" "ENSMUSG00000000028:Cdc45" "ENSMUSG00000000031:H19" "ENSMUSG00000000037:Scml2" ...
  ..$ logFC     : num [1:646608] -1.29 -6.4 -5.59 -3.94 -1.66 -8.94 0.17 2.35 0.36 -4.75 ...
  ..$ logCPM    : num [1:646608] 7.58 3.05 1.96 2.25 6.02 9.37 9.67 4.57 6.86 1.74 ...
  ..$ PValue    : num [1:646608] 0.158 0.066 0.0293 0.0484 0.31 ...
  ..$ FDR       : num [1:646608] 0.297 0.189 0.132 0.167 0.459 ...
  ..$ CellType  : chr [1:646608] "ShamMI" "ShamMI" "ShamMI" "ShamMI" ...
  ..$ Comparison: chr [1:646608] "B_cell-Rest" "B_cell-Rest" "B_cell-Rest" "B_cell-Rest" ...
}
\examples{
data(nMyo_Data)
## maybe str(nMyo_Data) ; plot(nMyo_Data) ...
}
\keyword{datasets}
