% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotSaver}
\alias{plotSaver}
\title{Plot generator function}
\usage{
plotSaver(myplot, type, output_folder, interactive, show.it, save.it, setDate)
}
\arguments{
\item{myplot}{data.frame. The plot to be generated.}

\item{type}{character. The type of plot.}

\item{output_folder}{character. The output folder.}

\item{interactive}{logical. If TRUE, the interative (plotly) plot is generated else a static (ggplot) is generated.}

\item{show.it}{logical. If TRUE, the plot is shown on screen.}

\item{save.it}{logical. If TRUE, the plot is stored in a file.}

\item{setDate}{character. The date stamp generated in CellsQuery().}
}
\value{
A data.frame.
}
\description{
Helper function
}
\keyword{plotSaver}
