% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ReadMarkers}
\alias{ReadMarkers}
\title{Read the Markers file}
\usage{
ReadMarkers(Markers_file)
}
\arguments{
\item{Markers_file}{character. The location of the markers file (it can be NULL).}
}
\value{
A vector of gene names or NULL
}
\description{
Helper function
}
\keyword{ReadMarkers}
