% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{RequiredDataReader}
\alias{RequiredDataReader}
\title{Read the required data files}
\usage{
RequiredDataReader(data_file)
}
\arguments{
\item{data_file}{character. The data file containing the normalised and corrected counts (CorrCounts), the design table (DesignTable) and the differential expressin analysis (DEs).}
}
\value{
A list of data
}
\description{
Helper function
}
\keyword{RequiredDataReader}
