% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Scatter}
\alias{Scatter}
\title{Generate the scatterplot}
\usage{
Scatter(
  Data,
  feature,
  grouping.by,
  highlight.by,
  showDims,
  dot.size,
  interactive
)
}
\arguments{
\item{Data}{data.frame. The data to use for plotting.}

\item{feature}{character. The feature to be plotted.}

\item{grouping.by}{character.  The grouping variable to be highlighted automatically.}

\item{highlight.by}{character.  The second grouping variable to be highlighted on click.}

\item{showDims}{character. The dimensions to be plotted.}

\item{dot.size}{numeric. The size of the dots to be plotted.}

\item{interactive}{logical. If TRUE, an interactive (plotly) plot is generated.}
}
\value{
Data for plotting
}
\description{
The function that generates the scatterplot.
}
\keyword{Scatter}
