% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ShapeSize}
\alias{ShapeSize}
\title{Assign dot shapes and sizes to the scatterplot}
\usage{
ShapeSize(plotData, my_colors, dot.size, feature, grouping.by)
}
\arguments{
\item{plotData}{data.frame. The data to use for plotting.}

\item{my_colors}{list. The outcome of makecolors_scatterplot().}

\item{dot.size}{numeric. The size of the dots to be plotted.}

\item{feature}{character. The feature to be plotted.}

\item{grouping.by}{character. The grouping variable to be highlighted.}
}
\value{
A data.frame
}
\description{
Helper function
}
\keyword{ShapeSize}
