% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tableSaver}
\alias{tableSaver}
\title{Table generator function}
\usage{
tableSaver(Data, save.table)
}
\arguments{
\item{Data}{data.frame. The outcome og MarkerQuery().}

\item{save.table}{logical. If TRUE it stores the table in a file.}
}
\value{
A list of tables
}
\description{
Helper function
}
\keyword{tableSaver}
