% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{testDataValidity}
\alias{testDataValidity}
\title{Read the DE_file and tests the data validity}
\usage{
testDataValidity(data)
}
\arguments{
\item{data}{list. The outcome of the RequiredDataReader() internal function.}
}
\value{
A list of data
}
\description{
Helper function
}
\keyword{testDataValidity}
