% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{testFilters}
\alias{testFilters}
\title{Check the filtering variables}
\usage{
testFilters(filters, dataMatrix, mode)
}
\arguments{
\item{filters}{character. The filters expression.}

\item{dataMatrix}{data.frame. The data.frame whose column names the filters refer to.}

\item{mode}{character. The function that the filter is applied to.}
}
\value{
Filters and possible mistakes in the filter names
}
\description{
Helper function
}
\keyword{testFilters}
