% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{testFiltersNumeric}
\alias{testFiltersNumeric}
\title{Check the numeric filtering variables}
\usage{
testFiltersNumeric(filters, dataMatrix, mode)
}
\arguments{
\item{filters}{character. The filters expression.}

\item{dataMatrix}{data.frame. The data.frame whose column names the filters refer to.}

\item{mode}{character. The function that the filter is applied to.}
}
\value{
Filters and possible mistakes in the filter names
}
\description{
Helper function
}
\keyword{testFiltersNumeric}
