% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validity}
\alias{validity}
\title{Test the data validity}
\usage{
validity(data)
}
\arguments{
\item{data}{list. It must include four components: component 'Counts' includes the normalised count data, component 'Design' includes the experimental design information, component 'Annotation' include the marker annotation (e.g. gene names and Ensembl IDs) and component 'DEstats' includes the differential expression analysis results.}
}
\description{
A helper (internal) function that reports an error if the data do not have the required format
}
\keyword{validity}
