% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Violin}
\alias{Violin}
\title{Generates the Violin plot}
\usage{
Violin(Data, feature, grouping.by, split.by, dot.size, interactive)
}
\arguments{
\item{Data}{list. The outcome of the MarkerQuery().}

\item{feature}{character. The feature to be plotted.}

\item{grouping.by}{character.  The grouping variable to be highlighted in the x-axis.}

\item{split.by}{character. The variable to split each violin with.}

\item{dot.size}{numeric. The size of the dots to be plotted.}

\item{interactive}{logical. If TRUE, an interactive (plotly) plot is generated.}
}
\value{
Data for plotting
}
\description{
The function that generates the violin plot
}
\keyword{Violin}
