% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Volcano}
\alias{Volcano}
\title{Generates the Volcano plot}
\usage{
Volcano(
  Data,
  de.comparison,
  logFC_cut,
  FDR_cut,
  show.more,
  dot.size,
  interactive
)
}
\arguments{
\item{Data}{list. The outcome of the MarkerQuery().}

\item{de.comparison}{list. A filter specifying the comparison to be plotted.}

\item{logFC_cut}{numeric. A positive value of the logFC cut-off to be depicted.}

\item{FDR_cut}{numeric. A value in (0,1] of the FDR cut-off to be depicted.}

\item{show.more}{logical. If TRUE, an extensive set of gene names are shown.}

\item{dot.size}{numeric. The size of the dots to be plotted.}

\item{interactive}{logical. If TRUE, an interactive (plotly) plot is generated.}
}
\value{
Data for plotting
}
\description{
The function that generates the volcano plot
}
\keyword{Volcano}
